python3 integration finished

+++ code/znet/src/znet_generator/layers/conv.py
@@ -1,5 +1,5 @@
 import copy
-from common import round_to_simd, generate_param_string, fill_tensor, \
+from .common import round_to_simd, generate_param_string, fill_tensor, \
                    zero_out_tensor, get_simd_width
 import numpy as np
 
@@ -20,8 +20,8 @@ def set_conv_dim(params, bot_tensor):
     top_dim[0] = bot_tensor.dim[0]
     top_dim[1] = params["ofm"]
     for i in [2, 3, 4]:
-	top_dim[i] = (bot_tensor.dim[i] - params["kernel_dim"][i] +
-		                         2*params["pad"][i - 2]) / params["stride"][i - 2] + 1
+        top_dim[i] = int((bot_tensor.dim[i] - params["kernel_dim"][i] +
+                                         2*params["pad"][i - 2]) / params["stride"][i - 2] + 1)
     params["top_dim"] = top_dim
     params["bot_dim"] = bot_tensor.dim
     params["bot_size"] = bot_tensor.memory_size
@@ -71,9 +71,9 @@ def block_kernel(kernel, lparam):
         total_ofms = round_to_simd(kdim[0], lparam["arch"])
         total_ifms = round_to_simd(kdim[1], lparam["arch"])
 
-        offset = ofm/S
-        offset *= total_ifms/S
-        offset += ifm/S
+        offset = int(ofm/S)
+        offset *= int(total_ifms/S)
+        offset += int(ifm/S)
         offset *= kdim[2]
         offset += kz
         offset *= kdim[3]
@@ -84,7 +84,7 @@ def block_kernel(kernel, lparam):
         offset += ifm % S
         offset *= S
         offset += ofm % S
-        return offset
+        return int(offset)
 
     # h5 weight format: ofm-ifm-kz-kx-ky
     # output format: ofm/S-ifm/S-kz-kx-ky-ofm%S-ifm%S
